Background and UI of the settings-activity
First commit, in which I added the background, consisting of lines, which can be permanent, or fading away (whih they can do inb three different ways). Also the first kind of clock was implemented, and the background-image is centered. The background does not yet take the settings (except the background-picture) from the user, which he can (also not yet) define in the settings-activity.