Fuck Android Studios Auto-completion!
Figured out the mistake with the clock was in the layout caused by android studios auto-completion suggesting me the wrong thing. In the seekbar instead of "max" I got "maxEMS", which is why there were only values upto 10% stored (The seekbars represent the percentage of the screen taken up by the clock).